+++ packages/analytics-react-native/test/config/joined-config.test.ts
@@ -34,8 +34,9 @@ describe('joined-config', () => {
       expect(localConfig.autocapture).toEqual({
         screenViews: false,
         sessions: true,
-        appState: false,
+        appLifecycles: false,
         elementInteractions: false,
+        networkTracking: false,
       });
     });
 
@@ -99,8 +100,9 @@ describe('joined-config', () => {
       const expectedAutocapture = {
         screenViews: true,
         sessions: false,
-        appState: true,
+        appLifecycles: true,
         elementInteractions: true,
+        networkTracking: true,
       };
 
       updateReactNativeConfigWithRemoteConfig(remoteConfig, localConfig);