+++ packages/segment-session-replay-plugin/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.38](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/segment-session-replay-plugin@0.0.37...@amplitude/segment-session-replay-plugin@0.0.38) (2026-07-22)
+
+**Note:** Version bump only for package @amplitude/segment-session-replay-plugin
+
+
+
+
+
+## [0.0.37](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/segment-session-replay-plugin@0.0.36...@amplitude/segment-session-replay-plugin@0.0.37) (2026-07-20)
+
+**Note:** Version bump only for package @amplitude/segment-session-replay-plugin
+
+
+
+
+
 ## [0.0.36](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/segment-session-replay-plugin@0.0.35...@amplitude/segment-session-replay-plugin@0.0.36) (2026-07-17)
 
 **Note:** Version bump only for package @amplitude/segment-session-replay-plugin